feat(mxfp4): fused dual-quant grouped GEMM with deterministic RHT + perf + tests
- Fuse grouped activation dual-quant: one bf16 read emits the tight-M rowwise
  fwd/dgrad operand and the 128-padded per-group colwise wgrad operand, with
  GPU-computed pad offsets (no bf16 scatter, no D2H sync -> CUDA-graph capturable).
- Batch the weight dual-quant over the 3D (G,N,K) weight in a single launch
  (blockIdx.z per group), removing the Python per-group loop and the full-weight
  transpose/contiguous copy; mirrors the MXFP8 weight path.
- Make rowwise-RHT quant bit-exact deterministic: fuse ds_swizzle + s_waitcnt
  lgkmcnt(0) into one memory-clobbering asm so the compiler cannot reorder the
  reduction across in-flight LDS stores and latch stale lane data.
- Skip the redundant per-pass warp_reduce_max_8 on the 2D-block scale path
  (tile-wide warp_reduce_max_64 already yields the same max; bit-exact, cuts
  VALU on the VALU-bound weight quant).
- Use fp32 accumulation in the grouped-GEMM reference to dodge a ROCm/hipBLASLt
  bf16 NT-GEMM fault on certain shapes.
- Expand the fp4 grouped-GEMM test suite to match the fp8 suite 1:1
  (mx_blockwise sweep, zero_group_lens, cuda_graph, deterministic) plus the
  fp4-specific unaligned-N/K padded-contraction case.

+++ tests/pytorch/ref/gemm_ref.py
@@ -11,12 +11,24 @@ import torch.nn as nn
 
 
 def grouped_gemm_ref(a, b, seg_lens, trans_b=True):
+    # NOTE: the per-group matmul is accumulated in fp32 (operands upcast, result
+    # cast back to ``a.dtype``) for two reasons:
+    #   1. Correctness: an fp32-accumulated golden is the right reference for a
+    #      low-precision (fp4/fp8) product comparison.
+    #   2. Robustness: on ROCm/hipBLASLt a *bf16* NT GEMM (row-major A @ a
+    #      transposed, column-major B) selects a kernel that reads out of bounds
+    #      for some odd group M (e.g. a 2078-row slice with N=2816, K=2048),
+    #      hard-faulting the GPU ("Memory access fault ... Reason: Unknown") in
+    #      both the fwd matmul and the autograd bwd (which re-introduces a
+    #      transposed bf16 operand). Running in fp32 dodges the bad bf16 kernel.
+    #      The product kernel under test runs on Triton, so only this reference
+    #      hit the library bug.
     seg_lens = seg_lens.cpu().numpy()
     out = []
     start = 0
     for i, size in enumerate(seg_lens):
         rhs = b[i, :, :].t() if trans_b else b[i, :, :]
-        out.append(a[start : start + size, :] @ rhs)
+        out.append((a[start : start + size, :].float() @ rhs.float()).to(a.dtype))
         start += size
     return torch.cat(out)
